+++ app/Main.hs
@@ -1,6 +1,24 @@
 module Main where
 
-import Lib
+import Compile
+import Text.Printf
+
+name = "AdventOfCode : 2021 : 24 - simpliest IR compiler"
+major_version = 0 :: Int
+minor_version = 1 :: Int
+
+input_file = "example_1/input.txt"
+output_prefix = "example_1/out"
+
+global_uses = ["globaluse z"]
 
 main :: IO ()
-main = someFunc
+main = do
+    text <- readFile input_file
+    let irs = compilerCompile [] global_uses text
+    compilerSavePhases output_prefix irs
+
+    putStrLn $ printf "%s. Version: %d.%d" name major_version minor_version
+    putStrLn $ printf "\tinput file = \"%s\"" input_file
+    putStrLn $ printf "\toutput files prefix = \"%s...\"" output_prefix
+    putStrLn "... [OK]"